refactor: switch from own our subrequest producer to graphql
