+++ modules/thunder_gqls/src/Plugin/GraphQL/DataProducer/ThunderBreadcrumb.php
@@ -1,0 +1,92 @@
+<?php
+
+namespace Drupal\thunder_gqls\Plugin\GraphQL\DataProducer;
+
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Drupal\graphql\GraphQL\Execution\FieldContext;
+use Drupal\Core\Routing\RouteMatchInterface;
+use Drupal\Core\Breadcrumb\BreadcrumbManager;
+use Drupal\Core\Cache\CacheableMetadata;
+
+/**
+ * Gets the breadcrumb of the current path.
+ *
+ * @DataProducer(
+ *   id = "thunder_breadcrumb",
+ *   name = @Translation("Breadcrumb"),
+ *   description = @Translation("Breadcrumb"),
+ *   produces = @ContextDefinition("any",
+ *     label = @Translation("Breadcrumb")
+ *   ),
+ *   consumes = {
+ *     "path" = @ContextDefinition("string",
+ *       label = @Translation("Path"),
+ *       required = TRUE
+ *     ),
+ *   }
+ * )
+ */
+class ThunderBreadcrumb extends ThunderEntitySubRequestBase implements ContainerFactoryPluginInterface {
+
+  /**
+   * The breadcrumb manager.
+   *
+   * @var \Drupal\Core\Breadcrumb\BreadcrumbBuilderInterface
+   */
+  protected $breadcrumbManager;
+
+  /**
+   * The route match service.
+   *
+   * @var \Drupal\Core\Routing\CurrentRouteMatch
+   */
+  protected $currentRouteMatch;
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $pluginId, $pluginDefinition) {
+    $producer = parent::create($container, $configuration, $pluginId, $pluginDefinition);
+    $producer->setCurrentRouteMatch($container->get('current_route_match'));
+    $producer->setBreadcrumbManager($container->get('breadcrumb'));
+    return $producer;
+  }
+
+  /**
+   * Sets the current route match.
+   *
+   * @param \Drupal\Core\Routing\RouteMatchInterface $currentRouteMatch
+   *   The current route match.
+   */
+  protected function setCurrentRouteMatch(RouteMatchInterface $currentRouteMatch): void {
+    $this->currentRouteMatch = $currentRouteMatch;
+  }
+
+  /**
+   * Sets the breadcrumb manager service.
+   *
+   * @param \Drupal\Core\Breadcrumb\BreadcrumbManager $breadcrumbManager
+   *   The breadcrumb manager service.
+   */
+  protected function setBreadcrumbManager(BreadcrumbManager $breadcrumbManager): void {
+    $this->breadcrumbManager = $breadcrumbManager;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function doResolve(CacheableMetadata $cacheableMetadata, FieldContext $fieldContext) {
+    $breadCrumb = [];
+    foreach ($this->breadcrumbManager->build(
+      $this->currentRouteMatch->getCurrentRouteMatch()
+    )->getLinks() as $link) {
+      $breadCrumb[] = [
+        'uri' => $link->getUrl()->toUriString(),
+        'title' => $link->getText(),
+      ];
+    }
+    return $breadCrumb;
+  }
+
+}
