+++ modules/thunder_gqls/src/Plugin/GraphQL/SchemaExtension/ThunderJsonLdSchemaExtension.php
@@ -23,9 +23,8 @@ class ThunderJsonLdSchemaExtension extends ThunderSchemaExtensionPluginBase {
     parent::registerResolvers($registry);
 
     $this->addFieldResolverIfNotExists('Query', 'jsonld',
-      $this->builder->produce('thunder_entity_sub_request')
-        ->map('path', $this->builder->fromArgument('path'))
-        ->map('key', $this->builder->fromValue('jsonld')
+      $this->builder->produce('thunder_jsonld')
+        ->map('path', $this->builder->fromArgument('path')
       )
     );
   }
