--- modules/thunder_gqls/thunder_gqls.services.yml
@@ -1,6 +1,0 @@
-services:
-  main_content_renderer.thunder_gqls:
-    class: Drupal\thunder_gqls\Render\MainContent\ThunderGqlsRenderer
-    arguments: ['@module_handler', '@renderer', '@breadcrumb', '@current_route_match', '@metatag.manager', '@language_manager']
-    tags:
-      - { name: render.main_content_renderer, format: thunder_gqls }
